requirements.txt
djangorestframework==3.15.1
Django==5.0.3
python-dotenv==1.0.1
psycopg2==2.9.9
psycopg==3.1.18
